feat: Session에 대한 DB 계층 구현 (테스트 미구현)

+++ src/main/resources/schema.sql
@@ -14,7 +14,7 @@ create table session (
     end_date timestamp not null,
     cover_image_id bigint not null,
     status varchar(30) not null,
-    price int,
+    price bigint,
     capacity int,
     created_at timestamp not null,
     updated_at timestamp,
